+++ packages/node/CHANGELOG.md
@@ -6,6 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.13.3] - 2022-11-09
+### Fixed
+- Fix issue with reindex missing bind (#1391)
+
+## [1.13.2] - 2022-11-08
+- Patch release with @subql/node-core fix
+
+## [1.13.1] - 2022-11-08
+- Patch release with @subql/node-core fix
+
+## [1.13.0] - 2022-11-07
+### Changed
+- Update polkadot to 9.7.1 (#1384)
+### Fixed
+- Fix logic with reindex and unfinalized height, also reset dynamic ds (#1382)
+- Improve dictionary query (#1371)
+
 ## [1.12.0] - 2022-10-28
 
 ### Added